Fixed trackpad scrolling to be buttery smooth at slow speeds on macOS (instead of being very discrete-seeming)

+++ cpp/open3d/visualization/rendering/MatrixInteractorLogic.h
@@ -74,7 +74,7 @@ public:
 
     /// Moves the matrix along the forward axis. (This is one type
     /// of zoom.)
-    virtual void Dolly(int dy, DragType drag_type);
+    virtual void Dolly(float dy, DragType drag_type);
     virtual void Dolly(float z_dist, Camera::Transform matrix);
 
 private:
@@ -93,7 +93,7 @@ protected:
     void SetMatrix(const Camera::Transform& matrix);
     float CalcRotateRadians(int dx, int dy);
     float CalcRotateZRadians(int dx, int dy);
-    float CalcDollyDist(int dy, DragType drag_type);
+    float CalcDollyDist(float dy, DragType drag_type);
 };
 
 }  // namespace rendering
